requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
beautifulsoup4==4.11.1; python_full_version >= '3.6.0'
bs4==0.0.1
certifi==2022.6.15; python_full_version >= '3.6.0'
charset-normalizer==2.1.0; python_full_version >= '3.6.0'
idna==3.3; python_version >= '3.5'
lxml==4.9.1
numpy==1.21.6; python_version < '3.10' and platform_machine != 'aarch64' and platform_machine != 'arm64'
pandas==1.3.5
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==0.20.0
pytz==2022.2.1
requests==2.28.1
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
soupsieve==2.3.2.post1; python_full_version >= '3.6.0'
urllib3==1.26.11; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5' and python_version < '4'
